backend/requirements.txt
# Core FastAPI and web framework
fastapi==0.116.1
uvicorn==0.35.0

# Database and ORM
sqlalchemy==2.0.43
alembic==1.16.5
psycopg2-binary==2.9.10

# Authentication and security
python-jose==3.3.0
passlib==1.7.4
bcrypt
python-multipart==0.0.20

# Environment and configuration
python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.10.1

# External services
openai==1.105.0
resend==2.13.1

# LangChain and AI (only what's actually used)
langchain==0.3.27
langchain-core==0.3.75
langchain-openai==0.3.32

# Utilities
email-validator==2.3.0

# Caching and Redis (only redis is used)
redis==5.2.0